refactor(user,common): colocate user and common test doubles beside their subjects
Move the remaining user and common mock factories out of the lib jest
mirror to sit next to the code they double: .mock.ts for jest.Mocked
stubs, .builder.ts for real-object builders (renamed to build*). The
common dir-barrel doubles sit as sibling .mock.ts files. This empties
the mirror, so lib/jest/ no longer exists.
