+++ src/common/application/auth-client.mock.ts
