+++ src/module/car/application/use-case/add.test.ts
@@ -6,8 +6,8 @@ import { createMockCarProvisioning } from '@/car/application/provisioning/car.mo
 import { AddCarUseCase } from '@/car/application/use-case/add';
 import type { AddCarApiRequest } from '@/car/interface/api/add.schema';
 import type { AuthClient } from '@/common/application/auth-client';
+import { createMockAuthClient } from '@/common/application/auth-client.mock';
 import { Result } from '@/common/application/result';
-import { createMockAuthClient } from '@/lib/jest/mock/src/common/application/auth-client';
 import { createMockAuthIdentity } from '@/test/mock/@supabase/auth';
 
 describe('AddCarUseCase', () => {
