+++ src/module/car/ownership/application/use-case/promote-primary-owner.test.ts
@@ -7,8 +7,8 @@ import { PromotePrimaryOwnerUseCase } from '@/car/ownership/application/use-case
 import { buildCarOwnership } from '@/car/ownership/domain/ownership/car-ownership.builder';
 import type { PromotePrimaryOwnerApiRequest } from '@/car/ownership/interface/api/promote.schema';
 import type { AuthClient } from '@/common/application/auth-client';
+import { createMockAuthClient } from '@/common/application/auth-client.mock';
 import { Result } from '@/common/application/result';
-import { createMockAuthClient } from '@/lib/jest/mock/src/common/application/auth-client';
 import { createMockAuthIdentity } from '@/test/mock/@supabase/auth';
 
 const PRIMARY_OWNER_ID = 'b5b55395-e32f-4376-be03-f66be0a63ec4';
