+++ src/module/car/ownership/application/use-case/remove-owner.test.ts
@@ -6,8 +6,8 @@ import { RemoveOwnerUseCase } from '@/car/ownership/application/use-case/remove-
 import { buildCarOwnership } from '@/car/ownership/domain/ownership/car-ownership.builder';
 import type { RemoveOwnerApiRequest } from '@/car/ownership/interface/api/remove.schema';
 import type { AuthClient } from '@/common/application/auth-client';
+import { createMockAuthClient } from '@/common/application/auth-client.mock';
 import { Result } from '@/common/application/result';
-import { createMockAuthClient } from '@/lib/jest/mock/src/common/application/auth-client';
 import { createMockAuthIdentity } from '@/test/mock/@supabase/auth';
 
 const PRIMARY_OWNER_ID = 'b5b55395-e32f-4376-be03-f66be0a63ec4';
