+++ src/module/car/ownership/infrastructure/data-source/ownership.test.ts
@@ -5,7 +5,7 @@ import { createMockOwnershipMapper } from '@/car/ownership/application/mapper/ow
 import { buildOwnershipPersistence } from '@/car/ownership/application/persistence-model/ownership.builder';
 import { Result } from '@/common/application/result';
 import type { SupabaseDatabaseClient } from '@/common/infrastructure/database-client/supabase';
-import { createMockSupabaseDatabaseClient } from '@/lib/jest/mock/src/common/infrastructure/supabase';
+import { createMockSupabaseDatabaseClient } from '@/common/infrastructure/database-client/supabase.mock';
 
 import { OwnershipDataSourceImplementation } from './ownership';
