+++ src/module/car/service-log/application/use-case/add-service-log.test.ts
@@ -9,8 +9,8 @@ import { createMockServiceLogRepository } from '@/car/service-log/application/re
 import { AddServiceLogUseCase } from '@/car/service-log/application/use-case/add-service-log';
 import type { AddServiceLogApiRequest } from '@/car/service-log/interface/api/add.schema';
 import type { AuthClient } from '@/common/application/auth-client';
+import { createMockAuthClient } from '@/common/application/auth-client.mock';
 import { Result } from '@/common/application/result';
-import { createMockAuthClient } from '@/lib/jest/mock/src/common/application/auth-client';
 import { createMockAuthIdentity } from '@/test/mock/@supabase/auth';
 
 const PRIMARY_OWNER_ID = 'b5b55395-e32f-4376-be03-f66be0a63ec4';
