+++ src/module/car/service-log/infrastructure/repository/service-log.test.ts
@@ -6,7 +6,7 @@ import { buildServiceLog } from '@/car/service-log/domain/service-log/service-lo
 import { ServiceLogRepositoryImplementation } from '@/car/service-log/infrastructure/repository/service-log';
 import { Result } from '@/common/application/result';
 import type { SupabaseDatabaseClient } from '@/common/infrastructure/database-client/supabase';
-import { createMockSupabaseDatabaseClient } from '@/lib/jest/mock/src/common/infrastructure/supabase';
+import { createMockSupabaseDatabaseClient } from '@/common/infrastructure/database-client/supabase.mock';
 
 describe('ServiceLogRepositoryImplementation', () => {
   let mockDbClient: jest.Mocked<SupabaseDatabaseClient>;
