+++ src/module/user/application/use-case/name-change.test.ts
@@ -1,14 +1,14 @@
 import type { AuthClient } from '@/common/application/auth-client';
+import { createMockAuthClient } from '@/common/application/auth-client.mock';
 import { Result } from '@/common/application/result';
-import { createMockAuthClient } from '@/lib/jest/mock/src/common/application/auth-client';
-import { createMockUserMapper } from '@/lib/jest/mock/src/module/user/application/mapper/user';
-import { createMockUserRepository } from '@/lib/jest/mock/src/module/user/application/user-repository';
-import { createMockUser } from '@/lib/jest/mock/src/module/user/domain/user/user';
 import { createMockAuthIdentity } from '@/test/mock/@supabase/auth';
 import type { UserDto } from '@/user/application/dto/user';
 import type { UserMapper } from '@/user/application/mapper/user';
+import { createMockUserMapper } from '@/user/application/mapper/user.mock';
 import type { UserRepository } from '@/user/application/repository/user';
+import { createMockUserRepository } from '@/user/application/repository/user.mock';
 import { NameChangeUseCase } from '@/user/application/use-case/name-change';
+import { buildUser } from '@/user/domain/user/user.builder';
 import type { NameChangeApiRequest } from '@/user/interface/api/name-change.schema';
 
 describe('NameChangeUseCase', () => {
@@ -41,7 +41,7 @@ describe('NameChangeUseCase', () => {
     };
 
     it('should change the user name successfully', async () => {
-      const mockUser = createMockUser();
+      const mockUser = buildUser();
 
       mockAuthClient.authenticate.mockResolvedValue(
         Result.ok(mockAuthIdentity),
@@ -112,7 +112,7 @@ describe('NameChangeUseCase', () => {
     });
 
     it('should fail as validation when name change validation fails', async () => {
-      const mockUser = createMockUser();
+      const mockUser = buildUser();
       const invalidContract: NameChangeApiRequest = {
         name: '',
       };
@@ -139,7 +139,7 @@ describe('NameChangeUseCase', () => {
     });
 
     it('should fail as unexpected when persistence fails', async () => {
-      const mockUser = createMockUser();
+      const mockUser = buildUser();
 
       mockAuthClient.authenticate.mockResolvedValue(
         Result.ok(mockAuthIdentity),
