+++ src/module/user/domain/user/user.builder.ts
@@ -3,7 +3,7 @@ import { Email } from '@/user/domain/user/value-object/email/email';
 import { Name } from '@/user/domain/user/value-object/name/name';
 import { UserId } from '@/user/domain/user/value-object/user-id/user-id';
 
-export function createMockUser({
+export function buildUser({
   id = '44dd8410-a912-480f-95be-9ad4cbe30d7f',
   email = 'test@example.com',
   name = 'Test User',
