+++ src/module/user/infrastructure/data-source/user.test.ts
@@ -1,14 +1,14 @@
 /* eslint testing-library/no-await-sync-queries:0 */
 import type { AuthClient } from '@/common/application/auth-client';
+import { createMockAuthClient } from '@/common/application/auth-client.mock';
 import { Result } from '@/common/application/result';
 import type { SupabaseDatabaseClient } from '@/common/infrastructure/database-client/supabase';
-import { createMockAuthClient } from '@/lib/jest/mock/src/common/application/auth-client';
-import { createMockSupabaseDatabaseClient } from '@/lib/jest/mock/src/common/infrastructure/supabase';
-import { createMockUserDto } from '@/lib/jest/mock/src/module/user/application/dto/user';
-import { createMockUserMapper } from '@/lib/jest/mock/src/module/user/application/mapper/user';
-import { createMockUserPersistence } from '@/lib/jest/mock/src/module/user/application/persistence-model/user';
+import { createMockSupabaseDatabaseClient } from '@/common/infrastructure/database-client/supabase.mock';
 import { createMockAuthIdentity } from '@/test/mock/@supabase/auth';
+import { buildUserDto } from '@/user/application/dto/user.builder';
 import type { UserMapper } from '@/user/application/mapper/user';
+import { createMockUserMapper } from '@/user/application/mapper/user.mock';
+import { buildUserPersistence } from '@/user/application/persistence-model/user.builder';
 
 import { UserDataSourceImplementation } from './user';
 
@@ -31,10 +31,10 @@ describe('UserDataSourceImplementation', () => {
 
   describe('getById', () => {
     const userId = '44dd8410-a912-480f-95be-9ad4cbe30d7f';
-    const userDto = createMockUserDto({ id: userId });
+    const userDto = buildUserDto({ id: userId });
 
     it('should return user DTO on success', async () => {
-      const persistence = createMockUserPersistence({ id: userId });
+      const persistence = buildUserPersistence({ id: userId });
 
       mockDbClient.query.mockResolvedValue(Result.ok(persistence));
       mockUserMapper.persistenceToDto.mockReturnValue(userDto);
@@ -71,23 +71,23 @@ describe('UserDataSourceImplementation', () => {
 
     it('should return array of user DTOs on success', async () => {
       const persistenceList = [
-        createMockUserPersistence({
+        buildUserPersistence({
           id: userIds[0],
           email: 'a@example.com',
           user_name: 'User A',
         }),
-        createMockUserPersistence({
+        buildUserPersistence({
           id: userIds[1],
           email: 'b@example.com',
           user_name: 'User B',
         }),
       ];
-      const dtoA = createMockUserDto({
+      const dtoA = buildUserDto({
         id: userIds[0],
         email: 'a@example.com',
         name: 'User A',
       });
-      const dtoB = createMockUserDto({
+      const dtoB = buildUserDto({
         id: userIds[1],
         email: 'b@example.com',
         name: 'User B',
@@ -123,10 +123,10 @@ describe('UserDataSourceImplementation', () => {
 
   describe('getSessionUser', () => {
     const authIdentity = createMockAuthIdentity();
-    const userDto = createMockUserDto({ id: authIdentity.id });
+    const userDto = buildUserDto({ id: authIdentity.id });
 
     it('should return user DTO on success', async () => {
-      const persistence = createMockUserPersistence({ id: authIdentity.id });
+      const persistence = buildUserPersistence({ id: authIdentity.id });
 
       mockAuthClient.authenticate.mockResolvedValue(Result.ok(authIdentity));
       mockDbClient.query.mockResolvedValue(Result.ok(persistence));
